requirements.txt
abel==0.0.6
asgiref==3.6.0
Babel==2.12.1
certifi==2020.6.20
chardet==5.1.0
charset-normalizer==3.1.0
defusedxml==0.6.0
Django==3.2.19
django-allauth==0.42.0
django-countries==7.5.1
django-crispy-forms==1.14.0
django-environ==0.10.0
django-extra-views==0.14.0
django-haystack==3.2.1
django-modeltranslation==0.15.1
django-multiselectfield==0.1.12
django-oscar==3.2
django-phonenumber-field==6.4.0
django-recaptcha==3.0.0
django-tables2==2.3.4
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
factory-boy==3.2.1
Faker==4.1.2
gunicorn==20.0.4
idna==2.10
# mysqlclient==2.1.1
oauthlib==3.2.2
phonenumbers==8.13.11
Pillow==9.5.0
purl==1.6
pycountry==22.3.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
razorpay==1.3.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
